pyproject.toml
[tool.poetry]
name = "gen-ai-application"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.9.7"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
streamlit = "^1.38.0"
openai = "^1.50.2"
pydantic-settings = "^2.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
